backend/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
sqlalchemy==1.4.36
python-dotenv==0.20.0
psycopg2
